Test Array iteration using `values`
Because the `@@iterator` property of the Array prototype is an alias for
`values`, the two methods can be used interchangeably. Using the
`values` property is preferable because it does not impose a requirement
on Symbols.

Re-write existing tests for iteration to use the `values` property.
Introduce a single test to ensure that the alias is defined properly.