Remove documentation on YAML syntax
Details on YAML may be helpful for some new contributors, but this
document should not attempt to cover the topic (especially not from the
description of a specific frontmatter entry). Replace with a link to a
more comprehensive source as this will be more generally useful to
those who need it (and less obtrusive for those who do not).